This Data analysis project include map chart which is available in 2016 or above version of Excel.
Important thing to know is while creating map plot, we must be careful about the country name and abbreviations we use must actually exist as this map plots are already available and we only have to fill the date by selecting the table. And also we use tabular data in value format to create map plot.